arm:misc: Change tfa9887 driver debug spew
Changed the driver for the tfa8997 device so that
it only prints a message for the SetParam function
if there is an error and audio is playing. If audio
is not playing access to the device is not possible
so an error is expected.

Bug 1233775

Change-Id: I5e3b4f7958104d28c74dac730ea936efed567325
Reviewed-on: http://git-master/r/198963

(cherry picked from commit 9bd34a41737e49321db885b48ee299976e1cb09f)

Signed-off-by: Scott Peterson <speterson@nvidia.com>
Change-Id: I5c3c3ad7296c1cd088030f2d6f007b33f04c3377
Reviewed-on: http://git-master/r/202731
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Anshul Jain (SW) <anshulj@nvidia.com>
GVS: Gerrit_Virtual_Submit