power: max17048: schedule initial work immediatly
Commit 985f1afc1757429de5d2d8e1d1589c1601717071 changed the delay
between two battery gauge checks, which results in the initial check
happening 30 seconds after boot. Because of this the Android framework
believes battery level is 0% when reaching the home screen and shuts the
device off if no external power is applied.

Fix this by scheduling the initial work immediatly so the correct
battery level is available when reaching the home screen.

Bug 1234222

Change-Id: Ie4d09d489dcd1266faa50f735dfa1e4ec9300505
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Reviewed-on: http://git-master/r/214841
Reviewed-by: Anshul Jain (SW) <anshulj@nvidia.com>