ARM: tegra11: clock: Complete host1x bus update algorithm
Finalized 1x bus rate update algorithm. It now guarantees to switch to
the target parent/rate without violation of max clock limits for any
combination of pll rates and max limits (not just for current Tegra11
settings). It would also attempt to switch without dip in bus rate if
it is possible, but this cannot be guaranteed (example: switch from
408 MHz : 1 to 624 MHz : 2 with maximum 1x bus limit 408 MHz will be
executed as 408 => 204 => 312 MHz, and there is no way to avoid rate
dip in this case).

Change-Id: I073effbcd997cefc949c9f90465227b7026538d6
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/227327
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>
Signed-off-by: Joe Maples <joe@frap129.org>