board-molly: Remove sdmmc3 and hookup ubik spi + gpio
To test ubik on Dalmore, we're using the external
SDMMC3 slot (which are the same pins we plan on using
in a realy Molly).  Disable the SDMMC3 driver.  Set
pinmux for the SPI3 and GPIOs we need.

Also disable config for SPI_SLAVE_TEGRA.  We have no
need to be a spi slave.

Enable the power supplies for SDMMC3 (there are two
of them on Dalmore) that we need for ubik:
 (1) VDD_3V3_SD to provide power to the module.
     This is LDO6 of TPS65090
 (2) VDDIO_SMMC3_AP to provide power to the
     SDMMC3 block of the Tegra.  This is LDO9 of
     palmas TPS65913 smps9

We mark them as always_on in board-molly-power.c
rather than add platform hooks for enabling them
at runtime.  Ubik really should be always on for
this product.

Change-Id: I7c2ff2f0542dd8b30aec5c5bbb853679a6527721
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>