Adjust BTLE settings to improve WiFi performance.
BTLE and WiFI share the same radio resources so they can interfere
with each other. We are experimenting with various settings
to see if we can improve WiFi performance when signal strength is low.

Changed to using 1 retry when BTLE TX fails. This requires a
custom driver from Marvell. Otherwise the setting of retries is ignored.

The settings that we are changing to improve WiFi also have an effect on
spontaneous disconnects between Wolfie and Bemote. Timeout was increased
to reduce spontaneous disconnects.

Timeout calculation modified to round up.

Bug: 9051165
Bug: 9421654
Change-Id: Ib93f27ef1dd7657cd5a2e73c02e6fb8abf26dde0
Signed-off-by: Phil Burk <philburk@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>