[kernel][btle] Low BTLE latency for early Bemotes
If the BTLE slave latency is high then early versions of Bemote
can timeout during encryption phase when connecting.
This change makes us use low latency until Bemote is running
a later version of the firmware.
In the later version, latency is ignored when Bemote is booted
and when it wakes up from hibernation.

We want this change in Bemote:
I62f570591db2fc64b4416fc3bd46861267e9924d

We know we have it if we have the version number set in this CL:
I2f55a45b427fca83984f058e7bf19ff521fdfcac

Bug: 10198360
Change-Id: Ib60123ce8c24efe8e088a21e6d29bc3b63bc2dfe
Signed-off-by: Phil Burk <philburk@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>