AahBtRemote: Fix crash bug with timeout timer
Delete the timer when exiting the le thread.  Otherwise, there
could be a race where the thread exists, and the timer structure
which is on the thread stack is still referenced when the timer
fires.

Don't delete the devices if the host setup fails.  The one-time
init of the devices is now never undone once they succeed, regardless
of whether host setup succeeds or not.  Host setup happens on every
start of the thread, and if it happens to fail on one of those starts,
there's not any benefit for tearing down the devices.

Change-Id: I621e0efb8e15ca353011ec15cec2e9a7ed3a3221
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>