AahBtRemote: Change some synchronization code
Rename some variables to make things clearer (at least to me):

tx_credit is a semaphore because it reflects the credit count
that we've hijacked for the le stack.  We reserve one credit
for the le stack and tx_credit reflects the usage of that one
credit.

cmd_wait is changed to a completion and renamed cmd_response.
it's used to wait for the completion of the response.

athome_bt_simple_cmd() is changed as follows:
* cmd_send semaphore is removed.  since athome_bt_simple_cmd()
  is only called from the le stack thread, synchronization isn't
  needed.  it's purely single threaded.
* the copying of the response is done in the response handler.
  this saves an extra copy.  it has one negative in that a
  bad dataOut ptr to athome_bt_simple_cmd() means that the
  crash will happen in another thread than in this one, but
  at this point, such bugs should be rare.

Change-Id: I7601f628470129fd40a212a2ece997a1949dff9e
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>