AahBtRemote: Fix a bug with sending two entropy packets
There's currently a race where Bemote service is requesting
an entropy packet to be sent to a bemote that's half bound
while at the same time the stack itself sends the encrypt
packet.  This winds up causing the Bemote to abort the
encryption process completely because it doesn't expect
double entropy packets.  Sometimes the race causes the
Bemote to respond with one entropy but the driver replaces
the entropy with a new one.

Fix by adding additional check to the userland request for
entropy/encryption by making sure the remote isn't half paired
since the encryption in that case is handled by the driver
itself.  Also change add_dev to not set bind_mode when
the LTK is added during pairing of a new remote since that
would cause the new check to take the wrong path.
This could change in the near future during Bemote
service redesign.

Change-Id: I34350ba02052c54cf69ac912a93396614fc252bd
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>