AahBtRemote: aahlog improvements
Remove aahlog_continue() and generally replace with aahlog_bytes()
or printing to a buffer and then using aahlog().

The problem with using aahlog_continue() is than when you turn on
extensive logging, the multi-threaded nature of the driver makes
multi-printk calls hard to read because inevitably one thread's
sequence of calls gets interrupted by another thread.

Change-Id: I9fc9298602a28a092d09ab5da6d9e95df9752c2d
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>