AahBtRemote: Don't send packets directly to SDIO driver, use main bt thread
The LE function to send packets to the chip should enqueue it's packets
for the main bt thread to send and not send them directly because the
chip might be sleeping.  The main bt thread does the wakeup
as needed.

Also, it appears we need to delay our first packets or else we don't
get a response.  Not clear why.  Only theory is that the firmware
or the Marvell driver is still busy doing some things that isn't
ready for our commands/requests yet.

Change-Id: I975557b256f087faac7bb92d7fe5fbd00b672ef8
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>