video: tegra: Improve hdmi worker
It's now a full blown state machine.  Ported from similar work
we did for OMAP.  Handles TV's that occassionally drop HPD
briefly.

This also fixes a deadlock that could happen because the
old worker function grabbed the dc->lock, but then later
calls a function that grabs the fb lock.  FB_BLANK ioctl
grabs the same locks but in the oppossite order.  The
new worker thread does not grab the dc->lock.

Change-Id: I410e4caf510bedfd1effd625bcea40121dd67e8b
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>