video: tegra: fb: preserve x/yoffset in check_var
tegra_fb_check_var() has an optional filter hook that is used for HDMI
displays. This hook works with an instance of struct fb_videomode and a
conversion from fb_var_screeninfo to fb_videomode is thus performed
prior to calling it (and back again since the hook might modify the
mode).

The issue is that fb_videomode has no equivalent of the xoffset
and yoffset member variables of fb_var_screeninfo, and that they are set
to 0 by fb_videomode_to_var(). This prevents double-buffering to work
when HDMI is the primary display: calls to the FBIOPUT_VSCREENINFO ioctl
always end up displaying the first page, since yoffset is invariably
reset.

This patch fixes this problem by preserving xoffset and yoffset in local
variables prior to calling the filter hook, and restoring them after
converting the mode back to fb_var_screeninfo.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Signed-off-by: Joe Maples <joe@frap129.org>