[bemote] Once we are LL encrypted, we are no longer binding.
This is a band-aid for bug 10655331.  It seems to have been introduced with
Mike's band-aid for the double encrypt issue discovered a few weeks
ago.  Basically, the "bind mode" flag was not getting cleared in the
kernel after Mike's patch.  If a bemote lost power or rebooted after
pairing with Molly, and Molly had also not rebooted, Molly would
refuse to connect to the bemote because it was advertising like it was
already paired (it was), but Molly thought that it was still in the
middle of pairing (because the flag had not been cleared)

Now, we clear the bind_mode flag as soon as we reach the encrypted
state at the link level.  This is hacky, but the re-factor of the
service cleans all of this up anyway, so I'm not too worried.

Bug: 10655331
Signed-off-by: John Grossman <johngro@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>