[bemote] Reset input timeline at startup.
Handle the case where bluedroid goes down and resets the entire
bluetooth stack (including the @home filter layer).  Previously, this
would not reset any of the input subsystem's state.  If bemote
happended to reconnect to the host and was still sending a stream of
input events with valid deltas, the host would continue to use the old
baseline event time, which would make all of the events appear old to
the app level.

The fix is simple (and should have been implemented behavior all
along); basically, just because the delta is valid on a packet from
bemote does not mean that the baseline time stamp is valid.  Make
certain to reset the baseline any time the driver is reset, and make
sure to latch a new timestamp instead of propagate if either the delta
from bemote is invalid, or of the internal baseline is invalid.

Bug: 10893660
Signed-off-by: John Grossman <johngro@google.com>
Change-Id: I6f7a5967902746cb3ba7624c94e7717aa4ba2d1b
Signed-off-by: Joe Maples <joe@frap129.org>