ARM: tegra: disable IKCONFIG and make modules built-in
Bug 1343930

To be conform with CTS, we need to disable IKCONFIG and MODULES,
and make existing modules all built-in.

CONFIG_MODULES and relevant options are left untouched, these will
be turned OFF if it's a user build. This option is still necessary
for engineering build for out-of-tree modules and many test cases.

Change-Id: Icd858d5d707e66d4d9499de8ad358d06bc4e35d2
Signed-off-by: Eric Miao <eric.miao@nvidia.com>
Reviewed-on: http://git-master/r/281984
Reviewed-by: Eric Miao <emiao@nvidia.com>
Tested-by: Eric Miao <emiao@nvidia.com>
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>