tegra: dc: do not probe HPD in atomic context
tegra_dc_hpd() uses a potentially sleeping gpio_get() call to check HPD
status. Move the call to tegra_dc_hpd() out of the atomic context to
prevent a potential lock.

The spinlock we are getting away from is supposed to protect the
worker structure and has nothing to do with HPD itself. By probing
HPD status once the worker structure has been updated, we are sure to
not miss any HPD event.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Signed-off-by: Joe Maples <joe@frap129.org>