mach: tegra: Change tegra_ram_console_debug_reserve()
Allow caller to specifying start address.  A 0 value means that
the function will choose the start address itself.

It's often better to have a fixed address that is less likely
to change because the default algorithm is based on putting
the ram console at the end of RAM, but other carveouts go at
the end of RAM first and any changes to the size of those
blocks can cause the ram console to move if that default
algorithm is used.

Change-Id: I4b8f8d86276f6285e4a1c214b4d97d65f8d19306
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>