net: support marking accepting TCP sockets
When using mark-based routing, sockets returned from accept()
may need to be marked differently depending on the incoming
connection request.

This is the case, for example, if different socket marks identify
different networks: a listening socket may want to accept
connections from all networks, but each connection should be
marked with the network that the request came in on, so that
subsequent packets are sent on the correct network.

This patch adds a sysctl to mark TCP sockets based on the fwmark
of the incoming SYN packet. If enabled, and an unmarked socket
receives a SYN, then the SYN packet's fwmark is written to the
connection's inet_request_sock, and later written back to the
accepted socket when the connection is established.  If the
socket already has a nonzero mark, then the behaviour is the same
as it is today, i.e., the listening socket's fwmark is used.

Black-box tested using user-mode linux:

- IPv4/IPv6 SYN+ACK, FIN, etc. packets are routed based on the
  mark of the incoming SYN packet.
- The socket returned by accept() is marked with the mark of the
  incoming SYN packet.
- Tested with syncookies=1 and syncookies=2.

Bug 200034037
Bug 200062531
Bug 200095892

Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
(cherry picked from commit 4593f09b1f9939ec6ed2f8d7848def26b98c47ac)
Change-Id: I26bc1eceefd2c588d73b921865ab70e4645ade57
Reviewed-on: http://git-master/r/498738
(cherry picked from commit 1eff0016fcec30c661eab86761beab1724c74473)
Signed-off-by: Steve Kuo <stevek@nvidia.com>
Reviewed-on: http://git-master/r/664925
GVS: Gerrit_Virtual_Submit
Signed-off-by: Nine Feng <nfeng@nvidia.com>
Reviewed-on: http://git-master/r/730772
Reviewed-by: Ximing Chen <ximingc@nvidia.com>